01_automation_move_download_files/requirements.txt
colorama==0.4.6
coverage==7.6.9
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
schedule==1.2.2
typing==3.7.4.3
